--- internal/storage/storageutil/folder_attrs.go
@@ -1,42 +1,0 @@
-// Copyright 2024 Google Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package storageutil
-
-import (
-	"strings"
-
-	"cloud.google.com/go/storage/control/apiv2/controlpb"
-	"github.com/googlecloudplatform/gcsfuse/v2/internal/storage/gcs"
-)
-
-func ControlFolderAttrsToGCSFolder(bucketName string, attrs *controlpb.Folder) *gcs.Folder {
-	// Setting the parameters in Folder and doing conversions as necessary.
-	return &gcs.Folder{
-		Name:           findFolderName(bucketName, attrs.Name),
-		Metageneration: attrs.Metageneration,
-		UpdateTime:     attrs.GetUpdateTime().AsTime(),
-	}
-}
-
-// In HNS, folder paths returned by control client APIs are in the form of:
-//
-// projects/_/buckets/{bucket}/folders/{folder}
-// This method extracts the folder name from such a string.
-func findFolderName(bucketName, object string) string {
-	prefix := "projects/_/buckets/" + bucketName + "/folders/"
-	folderName := strings.TrimPrefix(object, prefix)
-
-	return folderName
-}
